requirements.txt
# Python environment: python39

# Core Python version
python==3.9.19

# Core dependencies
numpy==1.26.4
scipy==1.13.1
pandas==1.5.3
matplotlib==3.8.4
seaborn==0.13.2
networkx==3.2.1
requests==2.32.3

torch==2.3.0+cu121
torchvision==0.18.0
torchaudio==2.3.0+cu121
torch-geometric==2.6.1
torch-cluster==1.6.3+pt23cu121
torch-scatter==2.0.7
torch-sparse==0.6.18+pt23cu121
torch-spline-conv==1.2.2+pt23cu121
deepspeed==0.16.3
xformers==0.0.26.post1
flash-attn==2.7.4.post1
triton==2.3.0
vllm==0.4.3
vllm-flash-attn==2.5.8.post2
bitsandbytes==0.45.1
lightgbm==4.5.0
scikit-learn==1.4.2
sentence-transformers==3.0.1
transformers==4.49.0
evaluate==0.4.3
peft==0.3.0
accelerate==1.4.0
dgl==2.1.0
faiss-gpu==1.7.2
umls-graph==0.0.2
pyserini==0.22.1
opentsne==1.0.1
spacy==3.7.5
scispacy==0.5.4
en-core-sci-sm==0.5.4
spark-nlp==5.5.1
hanlp-common==0.0.20
hanlp-restful==0.0.23
datasets==2.19.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pandarallel==1.6.5
joblib==1.4.0
h5py
gensim==4.3.3
altair==5.5.0
streamlit==1.42.2
gradio==4.44.1
pydeck==0.9.1
ray==2.11.0
tqdm==4.66.2
numba==0.59.1
cloudpickle==3.0.0
sqlalchemy==2.0.29
pymysql==1.1.1
dataset==1.6.2
chromadb==0.5.5
boto3==1.35.10
botocore==1.35.10
openai==1.10.0
cohere==4.45
duckduckgo-search==5.3.0
fastapi==0.110.2
uvicorn==0.29.0
starlette==0.37.2
httpx==0.27.0
aiohttp==3.9.5
aiofiles==23.2.1
wandb==0.16.6
sentry-sdk==2.0.1
prometheus-client==0.20.0
tiktoken==0.7.0
regex==2024.4.16
pydantic==2.7.4
pyyaml==6.0.1
rich==13.7.1
fire==0.6.0
tenacity==8.2.3
typer==0.12.5
jupyter==1.0.0
ipython==8.18.1
ipywidgets==8.1.2
notebook==7.1.3
jupyterlab==4.1.6
# GPU Utils
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-nccl-cu12==2.20.5
nvidia-nvtx-cu12==12.1.105
nvidia-ml-py==12.555.43
pynvml==11.5.0
rdkit==2023.9.6
mne==1.7.0
sympy==1.12
numexpr
